requirements.txt
Flask
SQLAlchemy
pymysql
python-dotenv
schedule